+++ CHANGELOG.md
@@ -7,6 +7,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.3.1] - Unreleased
 ### Added
 - Allow to modify the style of the logos block.
+- Allow to modify the theme of the speakers block.
 - Allow to modify the theme of the menu.
 - Allow to modify the theme of the footer.
 - Allow to configure the default meta tags options in the CMS.