+++ CHANGELOG.md
@@ -17,6 +17,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 - Allow to modify the theme of the activities block.
 - Allow to modify the theme of the people block.
 - Allow to configure the default meta tags options in the CMS.
+- New option "Invisible" for separators.
 
 ### Fixed
 - Calendar header layout.