+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.4.0] - Unreleased
+## [0.4.0] - 2025-07-22
 ### Added
 - New global theme per page.
 - New variable `--color-highlight` to configure shadow backgrounds.
@@ -80,7 +80,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-07-17
 First version
 
-[0.3.1]: https://github.com/tarugoconf/TOP/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/tarugoconf/TOP/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/tarugoconf/TOP/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/tarugoconf/TOP/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/tarugoconf/TOP/compare/v0.1.4...v0.2.0